#24 - moved Payments integration tests to PaymentsIntegrationTest
merged duplicate tests
testGetPayment was identical
testCreatePayment was more complete in DynamicClientIntegrationTest (retained the DynamicClientIntegrationTest version)
