+++ src/test/java/nl/stil4m/mollie/concepts/PaymentsIntegrationTest.java
@@ -5,11 +5,15 @@ import static nl.stil4m.mollie.TestUtil.VALID_API_KEY;
 import static nl.stil4m.mollie.TestUtil.assertWithin;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.fail;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
 
 import org.junit.Before;
@@ -37,15 +41,6 @@ public class PaymentsIntegrationTest {
         Client client = new ClientBuilder().withApiKey(VALID_API_KEY).build();
         payments = client.payments();
         issuers = client.issuers();
-    }
-    
-    @Test
-    public void testCreatePayment() throws IOException, InterruptedException {
-        Date beforeTest = new Date();
-        
-        ResponseOrError<CreatedPayment> payment = payments.create(new CreatePayment(Optional.empty(), 1.00, "Some description", "http://example.com", Optional.empty(), null));
-
-        assertWithin(beforeTest, payment.getData().getCreatedDatetime(), new Date(), 5000L);
     }
 
     @Test
@@ -97,5 +92,164 @@ public class PaymentsIntegrationTest {
         } catch (IllegalArgumentException e) {
             assertThat(e.getMessage(), is("Payment id may not be null"));
         }
+    }
+    
+
+
+    @Test
+    public void testCreatePayment() throws IOException {
+        Date beforeTest = new Date();
+        Map<String, Object> meta = new HashMap<>();
+        meta.put("foo", "bar");
+
+        ResponseOrError<CreatedPayment> payment = payments.create(new CreatePayment(Optional.empty(), 1.00, "Some description", "http://example.com", Optional.empty(), meta));
+
+        CreatedPayment createdPayment = payment.getData();
+        assertWithin(beforeTest, createdPayment.getCreatedDatetime(), new Date(), 5000L);
+
+        assertThat(createdPayment.getMethod(), is(nullValue()));
+        assertThat(createdPayment.getAmount(), is(1.00));
+        assertThat(createdPayment.getDescription(), is("Some description"));
+        assertThat(createdPayment.getId(), is(notNullValue()));
+        assertThat(createdPayment.getDetails(), is(nullValue()));
+        assertThat(createdPayment.getLinks(), is(notNullValue()));
+        assertThat(createdPayment.getLinks().getPaymentUrl().matches("https://www.mollie.com/payscreen/select-method/[A-Za-z0-9]+"), is(true));
+        assertThat(createdPayment.getLinks().getRedirectUrl(), is("http://example.com"));
+        assertThat(createdPayment.getLinks().getWebhookUrl(), is("https://stil4m.github.io"));
+        assertThat(createdPayment.getMode(), is("test"));
+        assertThat(createdPayment.getStatus(), is("open"));
+        assertThat(createdPayment.getMetadata(), is(meta));
+    }
+
+    @Test
+    public void testCreatePaymentWithMethod() throws IOException {
+        Date beforeTest = new Date();
+        Map<String, Object> meta = new HashMap<>();
+        meta.put("foo", "bar");
+
+        ResponseOrError<CreatedPayment> payment = payments.create(new CreatePayment(Optional.of("ideal"), 1.00, "Some description", "http://example.com", Optional.empty(), meta));
+
+        CreatedPayment createdPayment = payment.getData();
+        assertWithin(beforeTest, createdPayment.getCreatedDatetime(), new Date(), 5000L);
+
+        assertThat(createdPayment.getMethod(), is("ideal"));
+        assertThat(createdPayment.getAmount(), is(1.00));
+        assertThat(createdPayment.getDescription(), is("Some description"));
+        assertThat(createdPayment.getId(), is(notNullValue()));
+        assertThat(createdPayment.getDetails(), is(nullValue()));
+        assertThat(createdPayment.getLinks(), is(notNullValue()));
+        assertThat(createdPayment.getLinks().getPaymentUrl().matches("https://www.mollie.com/paymentscreen/ideal/select-issuer/[A-Za-z0-9]+"), is(true));
+        assertThat(createdPayment.getLinks().getRedirectUrl(), is("http://example.com"));
+        assertThat(createdPayment.getLinks().getWebhookUrl(), is(is("https://stil4m.github.io")));
+        assertThat(createdPayment.getMode(), is("test"));
+        assertThat(createdPayment.getStatus(), is("open"));
+        assertThat(createdPayment.getMetadata(), is(meta));
+    }
+
+
+    @Test
+    public void testCreateAndGetPayment() throws IOException {
+        Map<String, Object> meta = new HashMap<>();
+        meta.put("foo", "bar");
+
+        ResponseOrError<CreatedPayment> createResponse = payments.create(new CreatePayment(Optional.empty(), 1.00, "Some description", "http://example.com", Optional.empty(), meta));
+        ResponseOrError<Payment> paymentResponse = payments.get(createResponse.getData().getId());
+        Payment payment = paymentResponse.getData();
+
+        assertThat(payment.getMethod(), is(nullValue()));
+        assertThat(payment.getAmount(), is(1.00));
+        assertThat(payment.getDescription(), is("Some description"));
+        assertThat(payment.getId(), is(notNullValue()));
+        assertThat(payment.getDetails(), is(nullValue()));
+        assertThat(payment.getLinks(), is(notNullValue()));
+        assertThat(payment.getLinks().getPaymentUrl().matches("https://www.mollie.com/payscreen/select-method/[A-Za-z0-9]+"), is(true));
+        assertThat(payment.getLinks().getRedirectUrl(), is("http://example.com"));
+        assertThat(payment.getLinks().getWebhookUrl(), is(is("https://stil4m.github.io")));
+        assertThat(payment.getMode(), is("test"));
+        assertThat(payment.getStatus(), is("open"));
+        assertThat(payment.getMetadata(), is(meta));
+
+    }
+
+    @Test
+    public void testCreateAndGetCreditCardPayment() throws IOException {
+        Map<String, Object> meta = new HashMap<>();
+        meta.put("foo", "bar");
+
+        ResponseOrError<CreatedPayment> createResponse = payments.create(new CreatePayment(Optional.of("creditcard"), 2.00, "Some credit card description", "http://example.com", Optional.empty(), meta));
+        ResponseOrError<Payment> paymentResponse = payments.get(createResponse.getData().getId());
+        Payment payment = paymentResponse.getData();
+
+        assertThat(payment.getMethod(), is("creditcard"));
+        assertThat(payment.getAmount(), is(2.00));
+        assertThat(payment.getDescription(), is("Some credit card description"));
+        assertThat(payment.getId(), is(notNullValue()));
+        assertThat(payment.getDetails(), is(notNullValue())); // feeRegion=other
+        assertThat(payment.getLinks(), is(notNullValue()));
+        assertThat(payment.getLinks().getPaymentUrl().matches("https://www.mollie.com/paymentscreen/creditcard/testmode/[A-Za-z0-9]+"), is(true));
+        assertThat(payment.getLinks().getRedirectUrl(), is("http://example.com"));
+        assertThat(payment.getLinks().getWebhookUrl(), is(is("https://stil4m.github.io")));
+        assertThat(payment.getMode(), is("test"));
+        assertThat(payment.getStatus(), is("open"));
+        assertThat(payment.getMetadata(), is(meta));
+
+    }
+
+    @Test
+    public void testGetPayments() throws IOException, URISyntaxException {
+        ResponseOrError<Page<Payment>> result = payments.list(Optional.empty(), Optional.empty());
+        assertThat(result.getSuccess(), is(true));
+        Page<Payment> page = result.getData();
+
+        assertThat(page.getTotalCount(), is(notNullValue()));
+        assertThat(page.getData().size(), is(Math.min(10, page.getTotalCount())));
+        assertThat(page.getOffset(), is(0));
+        assertThat(page.getCount(), is(10));
+
+        assertThat(page.getLinks().getPrevious().isPresent(), is(false));
+        assertThat(page.getLinks().getNext().isPresent(), is(true));
+        assertThat(page.getLinks().getLast().isPresent(), is(true));
+        assertThat(page.getLinks().getFirst().isPresent(), is(true));
+    }
+
+    @Test
+    public void testGetPaymentsWithOffsetAndCount() throws IOException, URISyntaxException {
+        ResponseOrError<Page<Payment>> result = payments.list(Optional.of(20), Optional.of(40));
+        assertThat(result.getSuccess(), is(true));
+        Page<Payment> page = result.getData();
+
+        assertThat(page.getTotalCount(), is(notNullValue()));
+        assertThat(page.getData().size(), is(Math.min(20, page.getTotalCount())));
+        assertThat(page.getOffset(), is(40));
+        assertThat(page.getCount(), is(20));
+
+        assertThat(page.getLinks().getPrevious().isPresent(), is(true));
+        assertThat(page.getLinks().getNext().isPresent(), is(true));
+        assertThat(page.getLinks().getLast().isPresent(), is(true));
+        assertThat(page.getLinks().getFirst().isPresent(), is(true));
+    }
+
+    @Test
+    public void testNextPayments() throws IOException, URISyntaxException {
+        ResponseOrError<Page<Payment>> result = payments.list(Optional.empty(), Optional.empty());
+        assertThat(result.getSuccess(), is(true));
+        Page<Payment> page = result.getData();
+
+        Page<Payment> nextPage = payments.next(page).getData();
+
+        assertThat(nextPage.getOffset(), is(10));
+        assertThat(nextPage.getCount(), is(10));
+    }
+
+    @Test
+    public void testPreviousPayments() throws IOException, URISyntaxException {
+        ResponseOrError<Page<Payment>> result = payments.list(Optional.of(20), Optional.of(40));
+        assertThat(result.getSuccess(), is(true));
+        Page<Payment> page = result.getData();
+
+        Page<Payment> previousPage = payments.previous(page).getData();
+
+        assertThat(previousPage.getOffset(), is(20));
+        assertThat(previousPage.getCount(), is(20));
     }
 }
